refactor: merge fraud tests with e2e-tests
